requirement.txt.txt
IMAPClient==2.3.1
pyzmail36==1.0.4
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.5.2
